requirements.txt
streamlit
pandas
folium
streamlit-folium
scikit-learn
requests
polyline
python-dotenv
prophet
plotly
openai
xgboost
numpy
datetime
scipy